Clean up + add meaningful tests
