+++ lib/build.gradle
@@ -2,13 +2,13 @@
 
 buildscript {
     ext {
-        kotlin_version = '1.0.3'
+        kotlinVersion = '1.0.6'
 
-        rxjava_version = '1.1.9'
-        java_websocket_version = '1.3.0'
-        gson_version = '2.7'
+        rxjavaVersion = '1.2.5'
+        javaWebsocketVersion = '1.3.0'
+        gsonVersion = '2.7'
 
-        junit_version = '4.12'
+        junitVersion = '4.12'
     }
 
     repositories {
@@ -16,8 +16,8 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.2.3'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath 'com.android.tools.build:gradle:2.3.0-beta2'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
