+++ lib/devtools-reduks/build.gradle
@@ -13,7 +13,7 @@ dependencies {
 
     compile 'com.github.beyondeye.reduks:reduks-core:2.0.0b6'
 
-    testCompile "junit:junit:$junit_version"
+    testCompile "junit:junit:$junitVersion"
 }
 
 sourceCompatibility = "1.6"
