+++ lib/devtools-reduks/src/test/kotlin/com/github/kittinunf/redux/devTools/reduks/ReduksDevToolsTest.kt
@@ -1,0 +1,70 @@
+package com.github.kittinunf.redux.devTools.reduks
+
+import com.beyondeye.reduks.Reducer
+import com.beyondeye.reduks.SimpleStore
+import com.beyondeye.reduks.create
+import com.beyondeye.reduks.subscribe
+import org.hamcrest.MatcherAssert.assertThat
+import org.junit.Test
+import java.util.concurrent.CountDownLatch
+import org.hamcrest.CoreMatchers.`is` as isEqualTo
+
+class ReduksDevToolsTest {
+
+    data class CounterState(val count: Int = 0)
+
+    sealed class CounterAction {
+        object Init
+        object Increment
+        object Decrement
+    }
+
+    fun counterReducer() = Reducer<CounterState> { state, action ->
+        when (action) {
+            is CounterAction.Init -> CounterState()
+            is CounterAction.Increment -> {
+                state.copy(count = state.count + 1)
+            }
+            is CounterAction.Decrement -> {
+                state.copy(count = state.count - 1)
+            }
+            else -> state
+        }
+    }
+
+    @Test
+    fun `apply devtools in to reduks store enhancer`() {
+        val countdown = CountDownLatch(1)
+
+        val store = SimpleStore.Creator<CounterState>()
+                .create(counterReducer(), CounterState(), devTools<CounterState>())
+
+        var count = 0
+        store.subscribe {
+            count++
+            when (count) {
+                1 -> assertThat(store.state.count, isEqualTo(0))
+                2 -> assertThat(store.state.count, isEqualTo(1))
+                3 -> assertThat(store.state.count, isEqualTo(0))
+                4 -> assertThat(store.state.count, isEqualTo(1))
+                5 -> assertThat(store.state.count, isEqualTo(2))
+                6 -> assertThat(store.state.count, isEqualTo(1))
+                7 -> {
+                    assertThat(store.state.count, isEqualTo(2))
+                    countdown.countDown()
+                }
+            }
+        }
+
+        store.dispatch(CounterAction.Init) // = 0
+        store.dispatch(CounterAction.Increment) //+1
+        store.dispatch(CounterAction.Decrement) //-1
+        store.dispatch(CounterAction.Increment) //+1
+        store.dispatch(CounterAction.Increment) //+1
+        store.dispatch(CounterAction.Decrement) //-1
+        store.dispatch(CounterAction.Increment) //+1
+
+        countdown.await()
+    }
+
+}
