+++ lib/devtools-redux-kotlin/src/main/kotlin/com/github/kittinunf/redux/devTools/redux_kotlin/ReduxKotlinDevTools.kt
@@ -44,11 +44,7 @@ fun <S> devTools(option: InstrumentOption = defaultOption()): Store.Enhancer<S> 
                         store.dispatch(DevToolsStateChangeAction)
                     }
                 }
-
             }
         }
-
     }
 }
-
-
