+++ lib/devtools-redux-kotlin/src/test/kotlin/com/github/kittinunf/redux/devTools/redux_kotlin/ReduxKotlinDevToolsTest.kt
@@ -1,17 +1,13 @@
-package com.github.kittinunf.redux.devTools.reduks
+package com.github.kittinunf.redux.devTools.redux_kotlin
 
-import com.beyondeye.reduks.Reducer
-import com.beyondeye.reduks.SimpleStore
-import com.beyondeye.reduks.create
-import com.beyondeye.reduks.subscribe
+import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Test
+import redux.api.Reducer
+import redux.createStore
 import java.util.concurrent.CountDownLatch
+import org.hamcrest.CoreMatchers.`is` as isEqualTo
 
-/**
- * Created by kittinunf on 8/26/16.
- */
-
-class DevToolsTest {
+class ReduxKotlinDevToolsTest {
 
     data class CounterState(val count: Int = 0)
 
@@ -35,11 +31,27 @@ class DevToolsTest {
     }
 
     @Test
-    fun `apply devtools in to reduks store enhancer`() {
+    fun `apply devtools in to redux-kotlin store enhancer`() {
         val countdown = CountDownLatch(1)
 
-        val store = SimpleStore.Creator<CounterState>()
-                .create(counterReducer(), CounterState(), devTools<CounterState>())
+        val store = createStore(counterReducer(), CounterState(), devTools<CounterState>())
+
+        var count = 0
+        store.subscribe {
+            count++
+            when (count) {
+                1 -> assertThat(store.state.count, isEqualTo(0))
+                2 -> assertThat(store.state.count, isEqualTo(1))
+                3 -> assertThat(store.state.count, isEqualTo(0))
+                4 -> assertThat(store.state.count, isEqualTo(1))
+                5 -> assertThat(store.state.count, isEqualTo(2))
+                6 -> assertThat(store.state.count, isEqualTo(1))
+                7 -> {
+                    assertThat(store.state.count, isEqualTo(2))
+                    countdown.countDown()
+                }
+            }
+        }
 
         store.dispatch(CounterAction.Init) // = 0
         store.dispatch(CounterAction.Increment) //+1
@@ -48,10 +60,6 @@ class DevToolsTest {
         store.dispatch(CounterAction.Increment) //+1
         store.dispatch(CounterAction.Decrement) //-1
         store.dispatch(CounterAction.Increment) //+1
-
-        store.subscribe {
-            println(store.state)
-        }
 
         countdown.await()
     }
