+++ lib/devtools/src/main/kotlin/com/github/kittinunf/redux/devTools/core/Instrument.kt
@@ -2,12 +2,7 @@ package com.github.kittinunf.redux.devTools.core
 
 import com.github.kittinunf.redux.devTools.socket.SocketClient
 import com.google.gson.JsonParser
-import rx.subscriptions.CompositeSubscription
 import java.util.*
-
-/**
- * Created by kittinunf on 8/26/16.
- */
 
 data class InstrumentOption(val host: String, val port: Int, val name: String, val maxAge: Int)
 
@@ -21,11 +16,9 @@ class Instrument<S>(val options: InstrumentOption, val initialState: S) {
 
     private var currentStateIndex = -1
 
-    private val client: SocketClient
+    private val client: SocketClient = SocketClient(options.host, options.port)
 
     private val stateTimeLines = mutableListOf<S>()
-
-    private val subscriptionBag = CompositeSubscription()
 
     var onConnectionOpened: (() -> Unit)? = null
     var onMessageReceived: ((S) -> Unit)? = null
@@ -36,10 +29,6 @@ class Instrument<S>(val options: InstrumentOption, val initialState: S) {
         get() {
             return stateTimeLines.getOrNull(currentStateIndex) ?: initialState
         }
-
-    init {
-        client = SocketClient(options.host, options.port)
-    }
 
     fun start() {
         if (started) return
