+++ lib/devtools/src/main/kotlin/com/github/kittinunf/redux/devTools/core/InstrumentAction.kt
@@ -2,10 +2,6 @@ package com.github.kittinunf.redux.devTools.core
 
 import com.google.gson.JsonObject
 
-/**
- * Created by kittinunf on 9/5/16.
- */
-
 sealed class InstrumentAction(open val type: String, open val payload: Any? = null) {
 
     enum class ActionType {
